Artificial intelligence methods/photo quality improvement/requirements.txt
torchvision==0.20.1
torch==2.5.1
Pillow==11.0.0